+++ CHANGELOG.md
@@ -56,3 +56,6 @@ All notable changes to this project will be documented in this file, in reverse 
     displaying).
   - insert a closing `</span>` tag within the table data cell containing the
     required flag.
+- [#41](https://github.com/zfcampus/zf-apigility-documentation/pull/41) updates
+  the `ApiFactory` to ensure that if an entity has no collection associated with
+  it, documentation will not attempt to retrieve the fields.